fix: update test imports after moving inlet API route
- Update import paths from /apps/web/app/api/inlet to /app/api/inlet
- Fix TypeScript types in integration test event parsing
- All tests now pass (19 test suites, 118 tests)
